requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==2.15.6
PyYAML==5.4.1
requests==2.25.1
transbank-sdk @ git+https://github.com/TransbankDevelopers/transbank-sdk-python@a4212c725483dedd61057a86b92f544204391942
urllib3==1.26.3
Werkzeug==1.0.1
